add auth

+++ models/contacts.js
@@ -1,23 +1,23 @@
-const fs = require("fs/promises");
-const path = require("path");
 const ContactsModel = require("../db/models/contacts");
 
-const contactsPath = path.join(__dirname, ".", "contacts.json");
-const listContacts = async () => {
-  const fileData = await fs.readFile(contactsPath, (err) => {
-    if (err) throw err;
-  });
-  return JSON.parse(fileData);
+const listContacts = async ({ page, limit, favorite, userId }) => {
+  const skip = (page - 1) * limit;
+  const contactsQuery = ContactsModel.find({owner: userId})
+  if (favorite) {
+    contactsQuery.where("favorite").equals(favorite);
+  }
+  contactsQuery.skip(skip).limit(limit);
+  return contactsQuery;
 };
 
-const getContactById = async (contactId) => {
-  const contact = await ContactsModel.findById(contactId);
+const getContactById = async (contactId, userId) => {
+  const contact = await ContactsModel.findOne({ _id: contactId, owner: userId });
   return contact || null;
 };
 
-const removeContact = async (contactId) => {
+const removeContact = async (contactId, userId) => {
   try {
-    await ContactsModel.deleteOne({ _id: contactId });
+    await ContactsModel.deleteOne({ _id: contactId, owner: userId });
     return true;
   } catch (error) {
     console.error(error);
@@ -25,15 +25,15 @@ const removeContact = async (contactId) => {
   }
 };
 
-const addContact = async (body) => {
-  const contact = new ContactsModel(body);
+const addContact = async (body, userId) => {
+  const contact = new ContactsModel({...body, owner: userId});
   await contact.save();
   return contact;
 };
 
-const updateContact = async (contactId, body) => {
-  const updatedContact = await ContactsModel.findByIdAndUpdate(
-    contactId,
+const updateContact = async (contactId, body, userId) => {
+  const updatedContact = await ContactsModel.findOneAndUpdate(
+    { _id: contactId, owner: userId },
     {
       $set: {
         ...body,
